*** empty log message ***
git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@410 7152ba15-5bd8-0310-b828-e896570a432b

+++ APPLICATIONS/SimData/Include/SimData/ObjectInterface.h
@@ -187,7 +187,7 @@ public:
 	virtual void pack(Object *o, Packer &p) const {
 		C * object = dynamic_cast<C *>(o);
 		T &m = object->*member;
-		p.pack(m.size());
+		p.pack(static_cast<int>(m.size()));
 		typename T::iterator idx;
 		for (idx = m.begin(); idx != m.end(); idx++) {
 			p.pack(*idx);
